\name{is_in_track}
\alias{is_in_track}
\title{
Test whether points are in a track
}
\description{
Test whether points are in a track
}
\usage{
is_in_track(x, y, track_index = current_track_index())
}
\arguments{

  \item{x}{X-location of data points.}
  \item{y}{Y-location of data points.}
  \item{track_index}{Index of track.}

}
\value{
A logical vector.
}
\examples{
# There is no example
NULL

}
