\name{names.TRACK_META}
\alias{names.TRACK_META}
\title{
Names of all supported meta data
}
\description{
Names of all supported meta data
}
\usage{
\method{names}{TRACK_META}(x)
}
\arguments{

  \item{x}{Always use \code{TRACK_META}.}

}
\value{
A vector of characters.
}
\examples{
names(TRACK_META)
}
