\name{print.TRACK_META}
\alias{print.TRACK_META}
\title{
Print TRACK_META
}
\description{
Print TRACK_META
}
\usage{
\method{print}{TRACK_META}(x, ...)
}
\arguments{

  \item{x}{The \code{TRACK_META} object.}
  \item{...}{Additional parameters.}

}
\value{
No value is returned.
}
\examples{
# There is no example
NULL

}
