\name{spiral_clear}
\alias{spiral_clear}
\title{
Clear the spiral curve
}
\description{
Clear the spiral curve
}
\usage{
spiral_clear(check_vp = TRUE)
}
\arguments{

  \item{check_vp}{Whether to check the viewport.}

}
\details{
It basically sets the internally spiral object to NULL, and reset all the global options.
}
\value{
No value is returned.
}
\examples{
# There is no example
NULL

}
