\name{spiral_opt}
\alias{spiral_opt}
\title{
Global options
}
\description{
Global options
}
\usage{
spiral_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{

  \item{...}{Arguments for the parameters, see "details" section.}
  \item{RESET}{Whether to reset to default values.}
  \item{READ.ONLY}{Please ignore.}
  \item{LOCAL}{Please ignore.}
  \item{ADD}{Please ignore.}

}
\details{
There are following global parameters:

\describe{
  \item{\code{min_segment_len}}{Minimal length of the segment that partitions a curve.}
}
}
\examples{
# There is no example
NULL

}
