\name{polar_to_cartesian}
\alias{polar_to_cartesian}
\title{
Convert polar coordinates to catersian coordinates
}
\description{
Convert polar coordinates to catersian coordinates
}
\usage{
polar_to_cartesian(theta, r)
}
\arguments{

  \item{theta}{Angles, in radians.}
  \item{r}{Radius.}

}
\value{
A data frame with two columns: x abd y.
}
\examples{
# There is no example
NULL

}
